Changed names of files. The top Navbar script is now named navbar-layout.js and the Navbar itself is in navbar.js
