+++ src/components/index.js
@@ -1,3 +1,3 @@
 export { DashHeader as Header } from './header/header';
-export { Navbar } from './navbar/navbar';
+export { NavbarLayout } from './navbar/navbar-layout';
 export { RightSidebar } from './right-sidebar/right-sidebar';
